Allow plugins to decide the correct length for team packets
PacketEvents is forcing vanilla values on team packets. This causes the console to spam errors when a plugin tries to send team packets with increased values.

Example is 1.8 server w/1.21 clients. These clients can receive teams >16 thus TAB sends it. When packet events is installed, this behaviour spams errors.

Let plugins decide if it's fine or not to send those packets, and not have packet events cry over it.